pyproject.toml
[tool.poetry]
name = "repl_python3_Diimili"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.1"
emoji = "^1.4.2"
enos-mqtt-sdk-python = "^0.1.1"
python-dotenv = "^0.19.0"
scikit-image = "^0.18.3"
easy-pil = "^0.0.9"
Pillow-PIL = "^0.1dev"
pil-compat = "^1.0.0"
numpy1 = "^0.0.1"
pytesseract = "^0.3.8"
opencv-python = "^4.5.3"
youtube_dl = "^2021.6.6"
pafy = "^0.5.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
